Re-worked packaging in prep for an attempt to add to distributions.
It's not ready yet, but I might as well put some polish on this and
put it out into the cloud as a released to distributions solution.

As such, this adds packaging solutions for DEB and RPM based systems.
It's not yet ready for primetime, but it allows me to hand this
to some clients in a more ready-to-go manner right now.